requirements.txt
asgiref==3.8.1
Django==5.0.7
django-debug-toolbar==4.4.6
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
